requirements.txt
acres==0.3.0
aiofiles==24.1.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
cmake==4.0.0
configobj==5.0.9
configparser==7.2.0
cryptography==44.0.2
defusedxml==0.7.1
etelemetry==0.3.1
faiss-cpu==1.7.4
Faker==24.9.0
filelock==3.18.0
fitz==0.0.1.dev2
Flask==3.0.2
fonttools==4.56.0
fpdf2==2.7.7
frontend==0.0.3
fsspec==2025.3.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.23.0
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
lit==18.1.8
looseversion==1.3.0
lxml==5.3.1
MarkupSafe==3.0.2
mecab-python3==1.0.6
mpmath==1.3.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
prov==2.0.1
puremagic==1.28
pycparser==2.22
pydot==3.0.4
PyMuPDF==1.25.4
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==2.7.0
sentencepiece==0.2.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.3
threadpoolctl==3.6.0
tika==2.6.0
tokenizers==0.19.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.1
traits==7.0.2
transformers==4.41.2
triton==2.0.0
typing_extensions==4.13.0
tzdata==2025.2
unidic-lite==1.0.8
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
Whoosh==2.7.4
